### 2018-11-20 #1092 @arfon
:wave: @gidden: this PR fixes the YAML of your paper: https://github.com/IAMconsortium/pyam/pull/144
