### 2018-11-20 #1092 @gidden
Thanks again @arfon, PR is merge. This is my first submission to JOSS, will @whedon automatically rerender pdfs upon updates to `master` or do we need to explicitly ask it to do so?
