### 2018-11-20 #1092 @arfon
> we need to explicitly ask it to do so?

You'll need to request another build. You can do this with `@whedon generate pdf`
