### 2018-11-20 4b760fb
try fixing figure caption
