### 2018-11-22 #1092 @lheagy
Excellent, thanks @jtmiclat! Yes, starting on the weekend is fine
