### 2018-11-22 #1095 @whedon
[REVIEW]: pyam: a Python Package for the Analysis and Visualization of Assessment Models
**Submitting author:** @gidden (<a href="http://orcid.org/0000-0003-0687-414X">Matthew Gidden</a>)
**Repository:** <a href="https://github.com/IAMconsortium/pyam" target ="_blank">https://github.com/IAMconsortium/pyam</a>
**Version:** v0.1.2
