### 2018-11-26 f0f34ad
bump pyam paper title and refs
