+++ paper/paper.bib
@@ -44,16 +44,28 @@
 	pages = {3461--3482}
 }
 
-@article{huppmann_ncc_2018,
-	title = {A new scenario resource for integrated 1.5 °{C} research},
-	issn = {1758-6798},
-	url = {https://doi.org/10.1038/s41558-018-0317-4},
-	doi = {10.1038/s41558-018-0317-4},
-	abstract = {Scenarios have supported assessments of the IPCC for decades. A new scenario ensemble and a suite of visualization and analysis tools is now made available alongside the IPCC 1.5 °C Special Report to improve transparency and re-use of scenario data across research communities.},
-	journal = {Nature Climate Change},
-	author = {Huppmann, Daniel and Rogelj, Joeri and Kriegler, Elmar and Krey, Volker and Riahi, Keywan},
-	month = oct,
-	year = {2018}
+
+@article{riahi_shared_2017,
+	title = {The {Shared} {Socioeconomic} {Pathways} and their energy, land use, and greenhouse gas emissions implications: {An} overview},
+	volume = {42},
+	issn = {09593780},
+	shorttitle = {The {Shared} {Socioeconomic} {Pathways} and their energy, land use, and greenhouse gas emissions implications},
+	url = {http://linkinghub.elsevier.com/retrieve/pii/S0959378016300681},
+	doi = {10.1016/j.gloenvcha.2016.05.009},
+	language = {en},
+	urldate = {2017-04-26},
+	journal = {Global Environmental Change},
+	author = {Riahi, Keywan and van Vuuren, Detlef P. and Kriegler, Elmar and Edmonds, Jae and O’Neill, Brian C. and Fujimori, Shinichiro and Bauer, Nico and Calvin, Katherine and Dellink, Rob and Fricko, Oliver and Lutz, Wolfgang and Popp, Alexander and Cuaresma, Jesus Crespo and Kc, Samir and Leimbach, Marian and Jiang, Leiwen and Kram, Tom and Rao, Shilpa and Emmerling, Johannes and Ebi, Kristie and Hasegawa, Tomoko and Havlik, Petr and Humpenöder, Florian and Da Silva, Lara Aleluia and Smith, Steve and Stehfest, Elke and Bosetti, Valentina and Eom, Jiyong and Gernaat, David and Masui, Toshihiko and Rogelj, Joeri and Strefler, Jessica and Drouet, Laurent and Krey, Volker and Luderer, Gunnar and Harmsen, Mathijs and Takahashi, Kiyoshi and Baumstark, Lavinia and Doelman, Jonathan C. and Kainuma, Mikiko and Klimont, Zbigniew and Marangoni, Giacomo and Lotze-Campen, Hermann and Obersteiner, Michael and Tabeau, Andrzej and Tavoni, Massimo},
+	month = jan,
+	year = {2017},
+	pages = {153--168},
+	file = {1-s2.0-S0959378016300681-main.pdf:C\:\\Users\\gidden\\Zotero\\storage\\PIZSIHA3\\1-s2.0-S0959378016300681-main.pdf:application/pdf}
+}
+
+@book{rogelj_global_nodate,
+	edition = {In Press},
+	title = {Global warming of 1.5°{C}. {An} {IPCC} {Special} {Report} on the impacts of global warming of 1.5°{C} above pre-industrial levels and related global greenhouse gas emission pathways, in the context of strengthening the global response to the threat of climate change, sustainable development, and efforts to eradicate poverty},
+	author = {Rogelj, J. and Shindell, D. and Jiang, K. and Fifita, S. and Forster, P. and Ginzburg, V.}
 }
 
 @article{pyam_2018,
