### 2018-11-26 #1095 @gidden
Hi @lheagy, per your suggestion, I have updated the title to more accurately reflect the scientific domains in which `pyam` is or is envisioned being used. Looking forward to the reviews of @jtmiclat and @Chilipp.
