+++ pyam/__init__.py
@@ -18,3 +18,7 @@ try:
     warnings.formatwarning = custom_formatwarning
 except Exception:
     pass
+
+from ._version import get_versions
+__version__ = get_versions()['version']
+del get_versions
