+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include versioneer.py
+include pyam/_version.py
