+++ setup.cfg
@@ -1,2 +1,9 @@
 [bdist_wheel]
 universal=1
+
+[versioneer]
+VCS = git
+versionfile_source = pyam/_version.py
+versionfile_build = pyam/_version.py
+tag_prefix = v
+parentdir_prefix = pyam-
