### 2018-11-27 89c263f
update release notes
