### 2018-11-29 7a8b43b
reinsert NCC reference for IPCC assessment process and transparency aim
