### 2018-11-29 06b1eb5
Add extra tests and time filter (#9)
* Add test of extra col init behaviour

* Add failing tests of time filtering

* Setup time filtering tests

* Pass test filter year

* Redo tests of time filtering and include super messy first steps towards implementation

* Fill out tests and reset core

* Finish implementation of time filtering, cleaning up needed

* Refactor core so apply filters can use self.time_col
