+++ tests/conftest.py
@@ -150,9 +150,9 @@ CHECK_AGG_REGIONAL_DF = pd.DataFrame([
 
 TIME_AXES = [
     [2005, 2010],
-    [datetime(2005, 6, 17), datetime(2010, 6, 17)],
-    ['2005-06-17', '2010-06-17'],
-    ['2005-06-17 00:00:00', '2010-06-17 00:00:00']
+    [datetime(2005, 6, 17), datetime(2010, 7, 21)],
+    ['2005-06-17', '2010-07-21'],
+    ['2005-06-17 00:00:00', '2010-07-21 12:00:00']
 ]
 
 

[…]
