### 2018-11-29 e602b60
remove test for `meta` with extra columns (behaviour not supported)
