### 2018-11-29 fd39c71
reinsert NCC reference for IPCC assessment process and transparency aim (#146)
