+++ paper/paper.bib
@@ -84,3 +84,14 @@
 	 publisher={Zenodo},
      year = {2018}
 }
+
+@article{huppmann_ncc_2018,
+	title = {A new scenario resource for integrated 1.5 °{C} research},
+	issn = {1758-6798},
+	url = {https://doi.org/10.1038/s41558-018-0317-4},
+	doi = {10.1038/s41558-018-0317-4},
+	abstract = {Scenarios have supported assessments of the IPCC for decades. A new scenario ensemble and a suite of visualization and analysis tools is now made available alongside the IPCC 1.5 °C Special Report to improve transparency and re-use of scenario data across research communities.},
+	journal = {Nature Climate Change},
+	author = {Huppmann, Daniel and Rogelj, Joeri and Kriegler, Elmar and Krey, Volker and Riahi, Keywan},
+	month = oct,
+	year = {2018}
