### 2018-11-29 cfceab6
fix formatting of bib file for `pyam`paper
