+++ paper/paper.bib
@@ -95,3 +95,4 @@
 	author = {Huppmann, Daniel and Rogelj, Joeri and Kriegler, Elmar and Krey, Volker and Riahi, Keywan},
 	month = oct,
 	year = {2018}
+}
