### 2018-12-03 0faba75
update to style = pep440 and use the new command class
