+++ pyam/_version.py
@@ -40,7 +40,7 @@ def get_config():
     # _version.py
     cfg = VersioneerConfig()
     cfg.VCS = "git"
-    cfg.style = ""
+    cfg.style = "pep440"
     cfg.tag_prefix = "v"
     cfg.parentdir_prefix = "pyam-"
     cfg.versionfile_source = "pyam/_version.py"
