+++ setup.cfg
@@ -3,6 +3,7 @@ universal=1
 
 [versioneer]
 VCS = git
+style = pep440
 versionfile_source = pyam/_version.py
 versionfile_build = pyam/_version.py
 tag_prefix = v
