### 2018-12-03 6e8f9ed
add clean target to make file
