### 2018-12-03 2e2fe45
fix bug in error message match for Python 2
as suggested by @znicholls
