### 2018-12-03 30de073
update docstring for `filter()` and fix warning message formatting
