### 2018-12-03 d04891c
updating for versioneer (#145)
* updating for versioneer

* update release notes

* update to style = pep440 and use the new command class

* add clean target to make file

* dont remove vnev
