+++ setup.cfg
@@ -1,2 +1,10 @@
 [bdist_wheel]
 universal=1
+
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = pyam/_version.py
+versionfile_build = pyam/_version.py
+tag_prefix = v
+parentdir_prefix = pyam-
