+++ Makefile
@@ -1,3 +1,7 @@
+.PHONY: clean
+clean:
+	rm -rf build dist *egg-info
+
 .PHONY: new_release
 new_release:
 	@echo 'For a new release on PyPI:'
