### 2018-12-05 99e6132
test that to_excel writes a file
