+++ tests/test_core.py
@@ -49,6 +49,13 @@ def test_init_df_from_timeseries(test_df):
     pd.testing.assert_frame_equal(df.timeseries(), test_df.timeseries())
 
 
+def test_to_excel(test_df):
+    fname = 'foo_testing.xlsx'
+    test_df.to_excel(fname)
+    assert os.path.exists(fname)
+    os.remove(fname)
+
+
 def test_get_item(test_df):
     assert test_df['model'].unique() == ['a_model']
 
