### 2018-12-05 IAMconsortium/pyam#149 @gidden
Hey @Chilipp, thanks for raising this issue. I just tested on my linux machine, and all seems fine. Our CI tests on all three OSes, so I opened #151 to start to investigate.
