### 2018-12-05 #1095 @Chilipp
Dear @gidden and @danielhuppmann . I reviewed your software today and I very much appreciate your work. I find this as a very useful and nicely documented piece of software and I very much appreciated the review. There are only minor revisions that I can present

- [ ] **Installation instructions**
  > Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

  The list of dependencies is missing in the [install instructions](http://software.ene.iiasa.ac.at/pyam/install.html). From your setup.py I get, that pyam is built on

  - numpy
  - pandas >=0.21.0
  - matplotlib
  - seaborn

  They should at least be listed in the requirements, together with the necessary citations as specified on the webpages (e.g. for matplotlib it requires a link to https://matplotlib.org/ and a citation such as

  Hunter, J. D., Matplotlib: A 2D Graphics Environment, Computing in Science Engineering, 2007, 9, 90-95
  (see https://matplotlib.org/citing.html)
- [ ] **Automated tests**
  > Are there automated tests or manual steps described so that the function of the software can be verified?

  I highly recommend to include coverage tests into your software, e.g. by installing pytest-cov (pip install pytest-cov) on you CI services and then using some free services, such as [codecov.io (https://codecov.io/) or [coveralls.io](https://coveralls.io/). When I run your tests with coverage tests, I get

  <summary>pytest -v --mpl tests --cov=pyam</summary>

  ```
  Name                  Stmts   Miss  Cover
  -----------------------------------------
  pyam/__init__.py         18      0   100%
  pyam/_version.py        277    152    45%
  pyam/core.py            538     64    88%
  pyam/iiasa.py            92      3    97%
  pyam/logger.py            8      0   100%
  pyam/plotting.py        378    194    49%
  pyam/run_control.py      62     19    69%
  pyam/statistics.py      151     23    85%
  pyam/timeseries.py       53      2    96%
  pyam/utils.py           152     23    85%
  -----------------------------------------
  TOTAL                  1729    480    72%
  ```

  Overall, your package seems to be very well covered, but there are clearly some modules underrepresented in your tests (`pyam/plotting.py` and `pyam/_version.py`). However, JOSS does not formally require a report on your test coverage, so this is just a recommendation from my side.
- [ ] **Community guidelines:**
  > Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

  These are missing (see https://help.github.com/articles/setting-guidelines-for-repository-contributors/ for guidance)

Additionally there are some minor things that I added as issues and PRs to your project
- [ ] https://github.com/IAMconsortium/pyam/pull/147
- [ ] https://github.com/IAMconsortium/pyam/pull/148
- [ ] https://github.com/IAMconsortium/pyam/issues/149
- [ ] https://github.com/IAMconsortium/pyam/issues/150

Once all these items are checked off, I recommend to publish your paper.
