### 2018-12-05 0f82add
update scatter plot to support metadata coloring with variables
