### 2018-12-05 IAMconsortium/pyam#150 @gidden
Thanks for make this issue @Chilipp. It turns out I used a branch I was on for a presentation when I made this - PR #152 fixes it.
