+++ tests/test_core.py
@@ -58,7 +58,7 @@ def test_to_excel(test_df):
 
 def test_to_csv(test_df):
     fname = 'foo_testing.csv'
-    test_df.to_excel(fname)
+    test_df.to_csv(fname)
     assert os.path.exists(fname)
     os.remove(fname)
 
