### 2018-12-06 736d45e
test that to_excel writes a file (#151)
* test that to_excel writes a file

* also check csv

* fix function name in test

* try closing writer
