### 2018-12-06 IAMconsortium/pyam#149 @gidden
closed by #151
