### 2018-12-06 4427043
update release notes for #151
