### 2018-12-06 d76a3fa
Fix wrong links and RST (#147)
It seems like that the code in this file was marked with markdown style instead of RST. Furthermore, the links to `pyam.IamDataFrame.filter` and `pyam.filter_by_meta` where not correctly specified.
