### 2018-12-06 58d489c
update scatter plot to support metadata coloring with variables (#152)
* update scatter plot to support metadata coloring with variables

* update release notes

* add test
