### 2018-12-06 IAMconsortium/pyam#149 @Chilipp
All right, thanks. It worked for me as well.
