+++ .circleci/config.yml
@@ -16,20 +16,24 @@ jobs:
             mkdir /ci-artifacts;
             ls /ci-artifacts;
       - run:
-          name: Build Py2 Docker Image
-          command: /bin/sh -c "ln -f -s ci/py2/Dockerfile Dockerfile && docker build -t pyam/pyam-py2:latest ."
-          when: always
-      - run:
-          name: Python2 Tests
-          command:  docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py2:latest python2 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
-          when: always
-      - run:
           name: Build Py3 Docker Image
           command: /bin/sh -c "ln -f -s ci/py3/Dockerfile Dockerfile && docker build -t pyam/pyam-py3:latest ."
           when: always
       - run:
           name: Python3 Tests
-          command:  docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py3:latest python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
+          command: >
+            docker run -v /ci-artifacts:/test-artifacts -e COVERALLS_REPO_TOKEN pyam/pyam-py3:latest
+            /bin/bash -c "python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests --cov=pyam --cov-report term-missing && coveralls"
+          when: always
+      - run:
+          name: Build Py2 Docker Image
+          command: /bin/sh -c "ln -f -s ci/py2/Dockerfile Dockerfile && docker build -t pyam/pyam-py2:latest ."
+          when: always
+      - run:
+          name: Python2 Tests
+          command: >
+            docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py2:latest
+            python2 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
           when: always
       - run:
           name: Make Docs
