+++ .circleci/config.yml
@@ -22,8 +22,9 @@ jobs:
       - run:
           name: Python3 Tests
           command: >
+            echo $COVERALLS_REPO_TOKEN
             docker run -v /ci-artifacts:/test-artifacts -e COVERALLS_REPO_TOKEN pyam/pyam-py3:latest
-            /bin/bash -c "cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
+            /bin/bash -c "echo $COVERALLS_REPO_TOKEN && cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
           when: always
       - run:
           name: Build Py2 Docker Image
