+++ .circleci/config.yml
@@ -20,9 +20,11 @@ jobs:
           command: /bin/sh -c "ln -f -s ci/py3/Dockerfile Dockerfile && docker build -t pyam/pyam-py3:latest ."
           when: always
       - run:
+          name: Env Tests
+          command: echo $COVERALLS_REPO_TOKEN
+      - run:
           name: Python3 Tests
           command: >
-            echo $COVERALLS_REPO_TOKEN
             docker run -v /ci-artifacts:/test-artifacts -e COVERALLS_REPO_TOKEN pyam/pyam-py3:latest
             /bin/bash -c "echo $COVERALLS_REPO_TOKEN && cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
           when: always
