+++ .circleci/config.yml
@@ -4,6 +4,8 @@ jobs:
     working_directory: ~/pyam
     docker:
       - image: docker:17.05.0-ce-git
+        environment:
+          COVERALLS_REPO_TOKEN: $COVERALLS_REPO_TOKEN
     steps:
       - checkout
       - setup_remote_docker
