+++ .circleci/config.yml
@@ -5,7 +5,7 @@ jobs:
     docker:
       - image: docker:17.05.0-ce-git
         environment:
-          COVERALLS_REPO_TOKEN: $COVERALLS_REPO_TOKEN
+          COVERALLS_REPO_TOKEN: "$COVERALLS_REPO_TOKEN"
     steps:
       - checkout
       - setup_remote_docker
