### 2018-12-07 0d52356
rm printing of key
