+++ .circleci/config.yml
@@ -15,7 +15,6 @@ jobs:
       - run:
           name: Env Tests
           command: |
-            echo "COVERALLS_REPO_TOKEN $COVERALLS_REPO_TOKEN"; 
             echo "FOO_BAR $FOO_BAR";
             env
       - run:
