+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
             -e COVERALLS_REPO_TOKEN=$COVERALLS_REPO_TOKEN \
             -v /ci-artifacts:/test-artifacts \
             pyam/pyam-py3:latest \
-            /bin/bash -c "echo $COVERALLS_REPO_TOKEN && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing /pyam/tests && coveralls"
+            /bin/bash -c "echo $COVERALLS_REPO_TOKEN && cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
           when: always
       - run:
           name: Build Py2 Docker Image
