### 2018-12-17 09f24b3
first draft of contributor guidelines
