+++ CONTRIBUTING.rst
@@ -1,0 +1,22 @@
+Contributor Guidelines
+======================
+
+The pyam package has been developed with the explicit aim to facilitate
+open and collaborative analysis of integrated assessment and climate models.
+We appreciate contributions to the code base and development of new features.
+
+Please use the GitHub *Issues* feature to raise questions concerning potential
+bugs or to propose new features, but search and read resolved/closed topics on
+similar subjects before raising a new issue.
+
+For contributions to the code base, please use GitHub *Pull Requests*,
+including a detailed description of the new feature and unit tests
+to illustrate the intended functionality.
+Code submitted via pull requests must adhere to the `pep8`_ style formats.
+
+We do not require anyone to sign a *Contributor License Agreement*, because we
+believe that when posting ideas or submitting code to an open-source project,
+it should be obvious and self-evident that any such contributions
+are made in the spirit of open collaborative development.
+
+.. _`pep8`: https://www.python.org/dev/peps/pep-0008/
