+++ .circleci/config.yml
@@ -14,9 +14,17 @@ jobs:
             ls /ci-artifacts;
       - run:
           name: Env Tests
-          command: |
-            echo "FOO_BAR $FOO_BAR";
-            env
+          command: /bin/sh -c '[ -z "$COVERALLS_REPO_TOKEN" ] && echo "Need to set coveralls token" && exit 1;'
+      - run:
+          name: Build Py2 Docker Image
+          command: /bin/sh -c "ln -f -s ci/py2/Dockerfile Dockerfile && docker build -t pyam/pyam-py2:latest ."
+          when: always
+      - run:
+          name: Python2 Tests
+          command: >
+            docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py2:latest
+            python2 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
+          when: always
       - run:
           name: Build Py3 Docker Image
           command: /bin/sh -c "ln -f -s ci/py3/Dockerfile Dockerfile && docker build -t pyam/pyam-py3:latest ."
@@ -29,16 +37,6 @@ jobs:
             -v /ci-artifacts:/test-artifacts \
             pyam/pyam-py3:latest \
             /bin/bash -c "echo $COVERALLS_REPO_TOKEN && cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
-          when: always
-      - run:
-          name: Build Py2 Docker Image
-          command: /bin/sh -c "ln -f -s ci/py2/Dockerfile Dockerfile && docker build -t pyam/pyam-py2:latest ."
-          when: always
-      - run:
-          name: Python2 Tests
-          command: >
-            docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py2:latest
-            python2 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
           when: always
       - run:
           name: Make Docs
