### 2018-12-17 9bf9054
trying coveralls (#153)
add coveralls to ci
