### 2018-12-17 a5db2c4
add reference to googlegroups listserv
