+++ CONTRIBUTING.rst
@@ -5,6 +5,10 @@ The pyam package has been developed with the explicit aim to facilitate
 open and collaborative analysis of integrated assessment and climate models.
 We appreciate contributions to the code base and development of new features.
 
+The community mailing list for questions and discussions on new features is
+hosted by Googlegroups. Please join at `groups.google.com/d/forum/pyam`_
+and use <pyam@googlegroups.com> to send an email to the *pyam* community.
+
 Please use the GitHub *Issues* feature to raise questions concerning potential
 bugs or to propose new features, but search and read resolved/closed topics on
 similar subjects before raising a new issue.
@@ -18,5 +22,7 @@ We do not require anyone to sign a *Contributor License Agreement*, because we
 believe that when posting ideas or submitting code to an open-source project,
 it should be obvious and self-evident that any such contributions
 are made in the spirit of open collaborative development.
+
+.. _`groups.google.com/d/forum/pyam` : https://groups.google.com/d/forum/pyam
 
 .. _`pep8`: https://www.python.org/dev/peps/pep-0008/
