### 2018-12-17 971559e
add reference to googlegroup listserv on documentation page
