### 2018-12-17 4de3f13
Remove MANIFEST files
