+++ MANIFEST.IN
@@ -1,16 +1,0 @@
-# MANIFEST.in
-exclude .gitignore
-exclude appveyor.yml
-exclude .travis.yml
-recursive-exclude ci
-
-include README.md
-include setup.cfg
-include LICENSE
-
-prune .cache
-prune .git
-prune build
-prune dist
-
-recursive-exclude *.egg-info *
