+++ MANIFEST.in
@@ -1,2 +1,0 @@
-include versioneer.py
-include pyam/_version.py
