### 2018-12-17 4eb5351
Put MANIFEST back in
