### 2018-12-17 084854d
Remove duplicated line
