### 2018-12-18 d752623
Update README with development install structions too
