+++ README.md
@@ -72,3 +72,30 @@ To install from source after cloning this repository, simply run
 ```
 pip install -e .
 ```
+
+Development
++++++++++++
+
+To setup a development environment,
+
+```
+# pyam can be replaced with any other name
+# you don't have to specify your python version if you don't want
+conda create --name pip pyam python=X.Y.Z
+conda activate pyam  # may be source activate pyam or just activate pyam
+pip install -e .[tests,docs,deploy]
+# install other required packages (e.g. on a Unix like system)
+conda install -c conda-forge $(cat requirements.txt)
+# by hand also works e.g.
+conda install -c conda-forge cartopy geopandas
+# if you want to write notebooks
+pip install notebook nbconvert jupyter_contrib_nbextensions
+```
+
+To check everything has installed correctly,
+
+```
+pytest tests
+```
+
+All the tests should pass.
