### 2018-12-18 d59f451
Fix duplicate manifest files and update developer install instructions (#157)
* Remove MANIFEST files

* Put MANIFEST back in

* Remove duplicated line

* Update README with development install structions too
