+++ MANIFEST.IN
@@ -1,4 +1,3 @@
-# MANIFEST.in
 exclude .gitignore
 exclude appveyor.yml
 exclude .travis.yml
@@ -7,6 +6,8 @@ recursive-exclude ci
 include README.md
 include setup.cfg
 include LICENSE
+include versioneer.py
+include pyam/_version.py
 
 prune .cache
 prune .git
