### 2018-12-18 5ded261
Rework CI to use travis for all three OS's. This fixes issues with coveralls and removes immediately our circleci dep. Appveyor is still used as windows support on travis is still a bit wonky.
