+++ .travis.yml
@@ -1,25 +1,37 @@
-language: generic
+language: minimal
 
-os:
-  - osx
+matrix:
+    include:
+        - os: linux
+          env: PYENV=py37
+        - os: linux
+          env: PYENV=py27
+        - os: osx
+          env: PYENV=py37
+        - os: osx
+          env: PYENV=py27
+# turn these on once travis support gets a little better
+#        - os: windows
+#          env: PYENV=py37
+#        - os: windows
+#          env: PYENV=py27
 
 before_install:
-  # install conda
-  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh;
-  - chmod +x miniconda.sh
-  - ./miniconda.sh -b -p $HOME/miniconda
+  - bash ./ci/travis-install.sh
   - export PATH=$HOME/miniconda/bin:$PATH
-  - conda update --yes conda
-  # install deps
-  - conda install --yes cartopy geopandas matplotlib==2.1.2
-  - mkdir -p ~/.matplotlib
-  - touch ~/.matplotlib/matplotlibrc
-  - "echo 'backend: TkAgg' >> ~/.matplotlib/matplotlibrc"
+  - source activate pyam-testing
+  # This is required to get FreeType v2.6.1 instead of a newer version.
+  # If this does not happen, then slight changes occur in plots causing
+  # failures.
+  - conda uninstall --yes --force matplotlib
+  - pip install matplotlib==2.1.2
 
 install:
   - pip install -e .[tests]
-  - echo $PWD
+  - which coveralls
 
-# Run test
 script:
-  - pytest -v --mpl tests
+  - py.test tests --verbose --mpl --cov=pyam --cov-config ci/.coveragerc --cov-report term-missing
+
+after_success:
+  - coveralls
