+++ .circleci/config.yml
@@ -1,47 +1,0 @@
-version: 2
-jobs:
-  build:
-    working_directory: ~/pyam
-    docker:
-      - image: docker:17.05.0-ce-git
-    steps:
-      - checkout
-      - setup_remote_docker
-      - run:
-          name: CIRCLE_ARTIFACTS
-          command: |
-            mkdir /ci-artifacts;
-            ls /ci-artifacts;
-      - run:
-          name: Build Py2 Docker Image
-          command: /bin/sh -c "ln -f -s ci/py2/Dockerfile Dockerfile && docker build -t pyam/pyam-py2:latest ."
-          when: always
-      - run:
-          name: Python2 Tests
-          command: >
-            docker run -v /ci-artifacts:/test-artifacts pyam/pyam-py2:latest
-            python2 -m pytest -v --mpl --mpl-results-path=/test-artifacts /pyam/tests
-          when: always
-      - run:
-          name: Build Py3 Docker Image
-          command: /bin/sh -c "ln -f -s ci/py3/Dockerfile Dockerfile && docker build -t pyam/pyam-py3:latest ."
-          when: always
-      - run:
-          name: Python3 Tests
-          command: |
-            docker run \
-            -e COVERALLS_REPO_TOKEN=$COVERALLS_REPO_TOKEN \
-            -v /ci-artifacts:/test-artifacts \
-            pyam/pyam-py3:latest \
-            /bin/bash -c "echo $COVERALLS_REPO_TOKEN && cd /pyam/tests && python3 -m pytest -v --mpl --mpl-results-path=/test-artifacts --cov=pyam --cov-report term-missing && coveralls"
-          when: always
-      - run:
-          name: Make Docs
-          command:  docker run pyam/pyam-py3:latest /bin/bash -c "apt-get install --yes pandoc && cd /pyam/doc && pip install -r requirements.txt && make html"
-          when: always
-      - run:
-          name: check artifacts
-          command: ls /ci-artifacts
-          when: on_fail
-      - store_artifacts:
-          path: /ci-artifacts
