### 2018-12-18 99321d5
enable initializing with wide format and datetime columns
