### 2018-12-19 bb48706
Rework CI to use travis for all three OS's. This fixes issues with coveralls and removes immediately our circleci dep. Appveyor is still used as windows support on travis is still a bit wonky. (#158)
