+++ ci/py2/Dockerfile
@@ -1,5 +1,0 @@
-FROM gidden/python2-viz
-
-COPY . /pyam
-WORKDIR /
-RUN cd /pyam && python2 setup.py install 
