### 2018-12-19 6876b33
add contributor guidelines (#156)
* first draft of contributor guidelines

* add reference to googlegroups listserv

* add reference to googlegroup listserv on documentation page
