### 2018-12-19 f0da2ea
readd building docs to ci, closes #159
