+++ .travis.yml
@@ -28,10 +28,12 @@ before_install:
 
 install:
   - pip install -e .[tests]
-  - which coveralls
 
 script:
   - py.test tests --verbose --mpl --cov=pyam --cov-config ci/.coveragerc --cov-report term-missing
+  - cd doc
+  - pip install -r requirements.txt && make html
+  - cd ..
 
 after_success:
   - coveralls
