### 2018-12-19 0c862c1
add contributor to docs
