### 2018-12-19 a1b0163
move info to contribs
