+++ doc/source/index.rst
@@ -75,16 +75,6 @@ Documentation
    examples/index
    api
 
-
-Get in touch
-------------
-
-- Report bugs, suggest features or view the source code `on GitHub`_.
-- For less well defined questions or ideas, use the `mailing list`_.
-
-.. _mailing list: https://groups.google.com/forum/#!forum/pyam
-.. _on GitHub: http://github.com/IAMconsortium/pyam
-
 .. include:: ../../CONTRIBUTING.rst
 
 License
