+++ CONTRIBUTING.rst
@@ -1,13 +1,20 @@
-Contributor Guidelines
-======================
+
+Have a question? Get in touch!
+------------------------------
+
+- Report bugs, suggest features or view the source code `on GitHub`_.
+- For less well defined questions or ideas, use the `mailing list`_.
+
+.. _mailing list: https://groups.google.com/forum/#!forum/pyam
+.. _on GitHub: http://github.com/IAMconsortium/pyam
+
+
+Interested in Contributing? Welcome to the team!
+------------------------------------------------
 
 The pyam package has been developed with the explicit aim to facilitate
 open and collaborative analysis of integrated assessment and climate models.
 We appreciate contributions to the code base and development of new features.
-
-The community mailing list for questions and discussions on new features is
-hosted by Googlegroups. Please join at `groups.google.com/d/forum/pyam`_
-and use <pyam@googlegroups.com> to send an email to the *pyam* community.
 
 Please use the GitHub *Issues* feature to raise questions concerning potential
 bugs or to propose new features, but search and read resolved/closed topics on
@@ -22,7 +29,5 @@ We do not require anyone to sign a *Contributor License Agreement*, because we
 believe that when posting ideas or submitting code to an open-source project,
 it should be obvious and self-evident that any such contributions
 are made in the spirit of open collaborative development.
-
-.. _`groups.google.com/d/forum/pyam` : https://groups.google.com/d/forum/pyam
 
 .. _`pep8`: https://www.python.org/dev/peps/pep-0008/
