### 2018-12-19 4415a88
move tests related to aggregation checks to own test script
