### 2018-12-19 75f9374
move tests related to aggregation checks to own test script
