### 2018-12-19 ed7b032
move tests related to aggregation checks to own test script
