### 2018-12-19 b4b97e9
accept `pd.Series` as input
