### 2018-12-19 4cedadd
readd building docs to ci, closes #159 (#160)
