### 2018-12-19 43e0ae9
try to do install reqs
