+++ requirements.py
@@ -1,0 +1,12 @@
+install_requirements = [
+    "argparse",
+    "numpy",
+    "requests",
+    "pandas >=0.21.0",
+    "PyYAML",
+    "xlrd",
+    "xlsxwriter",
+    "matplotlib",
+    "seaborn",
+    "six",
+]
