### 2018-12-19 7e9eac5
make check-aggregates work including unit
