### 2018-12-19 95f1756
make check-aggregates work including unit
