### 2018-12-19 080db63
make check-aggregates work including unit
