### 2018-12-19 1d52d6a
add a function to aggregate components
