### 2018-12-19 dbefc91
Docclean (#161)
* add contributor to docs

* move info to contribs

* try to do install reqs

* appease stickler

* now print correctly

* add pdf to gitignore

* fill in references
