+++ doc/source/install.rst
@@ -27,6 +27,15 @@ From Source
 Depedencies
 ~~~~~~~~~~~
 
+Like any software project, we stand on the shoulders of giants. Our particular
+giants include :code:`numpy` :cite:`numpy`, :code:`matplotlib`
+:cite:`matplotlib`, and :code:`pandas` :cite:`pandas`. Explicit requirements are
+fully enumerated below.
+
+The required depedencies for :code:`pyam` are:
+
+  .. program-output:: python -c 'import sys; sys.path.append("../.."); import requirements; requirements.display()'
+
 The optional depedencies for :code:`pyam` are:
 
   .. include:: ../../requirements.txt
@@ -38,3 +47,10 @@ The depedencies for building this documentation are:
   .. include:: ../requirements.txt
 	  :start-line: 0
 	  :literal:
+
+References
+**********
+
+.. bibliography:: refs.bib
+   :style: plain 
+   :all:
