+++ doc/source/refs.bib
@@ -1,0 +1,33 @@
+
+@Misc{numpy,
+author =    {Travis Oliphant},
+title =     {{NumPy}: A guide to {NumPy}},
+year =      {2006--},
+howpublished = {USA: Trelgol Publishing},
+url = "http://www.numpy.org/"
+}
+
+@Article{matplotlib,
+Author    = {Hunter, J. D.},
+Title     = {Matplotlib: A 2D graphics environment},
+Journal   = {Computing In Science \& Engineering},
+Volume    = {9},
+Number    = {3},
+Pages     = {90--95},
+abstract  = {Matplotlib is a 2D graphics package used for Python
+for application development, interactive scripting, and
+publication-quality image generation across user
+interfaces and operating systems.},
+publisher = {IEEE COMPUTER SOC},
+doi       = {10.1109/MCSE.2007.55},
+year      = 2007
+}
+
+@InProceedings{pandas,
+  author    = {Wes McKinney},
+  title     = {Data Structures for Statistical Computing in Python},
+  booktitle = {Proceedings of the 9th Python in Science Conference},
+  pages     = {51 - 56},
+  year      = {2010},
+  editor    = {Stefan van der Walt and Jarrod Millman}
+}
