### 2018-12-19 8347f2f
add script to upload new website
