### 2018-12-19 #1095 @gidden
Hello all. First of all, thank you very much to @Chilipp and @jtmiclat for their very useful reviews. We have striven to implement all suggestions both of you made.

Notably these include:

- adding contributor/community guidance ([pr](https://github.com/IAMconsortium/pyam/pull/161), [site](https://software.ene.iiasa.ac.at/pyam/index.html#interested-in-contributing-welcome-to-the-team))
- adding coveralls support ([pr](https://github.com/IAMconsortium/pyam/pull/158), [site](https://coveralls.io/github/IAMconsortium/pyam))
- adding explicit required dependencies outside of `setup.py`/`pip` ([pr](https://github.com/IAMconsortium/pyam/pull/161), [site](https://software.ene.iiasa.ac.at/pyam/install.html#depedencies))

From our view, we believe the article is now ready for publication. Please let us know if we should do anything further.
