### 2018-12-19 66e1ae8
add to release notes
