### 2018-12-19 #1095 @Chilipp
I am okay with accepting his software for publication @lheagy. Thank you very much @gidden and @danielhuppmann for your work on this! 🎉 🎉
