### 2018-12-19 677bbbe
extend initialization of IamDataFrame for extra columns, distinguish `year` and `time`
