### 2018-12-19 0211359
refactor from static to `self._LONG_IDX`
