### 2018-12-19 085cbea
refactor `format_data()` to accept datetime as columns
