### 2018-12-19 4b78be9
add `iamc_index` as kwarg to `to_csv()` and `to_excel()`
