### 2018-12-19 072d975
refactor `_df` to `_data` in `__init__()`
