+++ tests/conftest.py
@@ -1,13 +1,12 @@
-import matplotlib
-matplotlib.use('agg')
-
 import os
 import pytest
-from datetime import datetime
-
 import pandas as pd
+import matplotlib
 
+from datetime import datetime
 from pyam import IamDataFrame
+
+matplotlib.use('agg')
 
 here = os.path.dirname(os.path.realpath(__file__))
 IMAGE_BASELINE_DIR = os.path.join(here, 'expected_figs')
