### 2018-12-19 df399d6
Add a function to aggregate timeseries components (#162)
* move tests related to aggregation checks to own test script

* accept `pd.Series` as input

* remove imports not required

* make check-aggregates work including unit

* add a function to aggregate components

* appease stickler

* add to release notes

* try to fix CI on travis
