### 2018-12-19 3a19b4e
Fix up README instructions
