### 2018-12-19 #1095 @lheagy
Many thanks @Chilipp and @jtmiclat for taking the time to review! 🎉

Congratulations @gidden on your submission! To finalize the acceptance, please archive your software on [zenodo](http://zenodo.org) or similar and post the doi here.
