### 2018-12-19 21805d8
Try putting matplotlib import stuff first
