### 2018-12-20 8a3b985
make stickler ignore E402
