### 2018-12-20 627a2b4
extend IamDataFrame to use extra data columns and sub-annual time (#167)
* extend unit tests for input as `datetime`

* refactor `timeseries()` to use `time_col`

* enable filtering by extra columns in `data`

* add check that no column conflicts exist between `meta` and `data`

* raise error when using `append()` with incompatible time formats

* docstring clean-up

* pep8

* test additional time formats

* clean up returned json object returned from IIASA db

* add kwarg `iamc_index` to `timeseries()` for clean or full index

* fix bug in setting `extra_cols` from long format

* when retriev
[…]
