+++ pyam/iiasa.py
@@ -12,7 +12,7 @@ except ImportError:
 
 from pyam.core import IamDataFrame
 from pyam.logger import logger
-from pyam.utils import LONG_IDX, isstr, pattern_match
+from pyam.utils import META_IDX, isstr, pattern_match
 
 # quiet this fool
 logging.getLogger('requests').setLevel(logging.WARNING)
@@ -176,7 +176,26 @@ class Connection(object):
         data = json.dumps(self._query_post_data(**kwargs))
         url = self.base_url + 'runs/bulk/ts'
         r = requests.post(url, headers=headers, data=data)
-        return pd.read_json(r.content, orient='records')
+        # refactor returned json object to be castable to an IamDataFrame
+        df = (
+            pd.read_json(r.content, orient='records')
+            .drop(columns='runId')
+            .rename(columns={'time': 'subannual'})
+        )
+        # check if returned dataframe has subannual disaggregation, drop if not
+        if pd.Series([i in [-1, 'year'] for i in df.subannual]).all():
+            df.drop(columns='subannual', inplace=True)
+        # check if there are multiple version for any model/scenario
+        lst = (
+            df[META_IDX + ['version']].drop_duplicates()
+            .groupby(META_IDX).count().version
+        )
+        if max(lst) > 1:
+            raise ValueError('multiple versions for {}'.format(
+                lst[lst > 1].index.to_list()))
+        df.drop(columns='version', inplace=True)
+
+        return df
 
 
 def read_iiasa(name, **kwargs):
@@ -185,7 +204,7 @@ def read_iiasa(name, **kwargs):
     """
     conn = Connection(name)
     df = conn.query(**kwargs)
-    return IamDataFrame(df[LONG_IDX + ['value']])
+    return IamDataFrame(df)
 
 
 def read_iiasa_iamc15(**kwargs):
