### 2018-12-27 #1095 @danielhuppmann
thanks @lheagy, @Chilipp and @jtmiclat for the useful feedback and positive reviews.

@gidden is on an extended Christmas break and I'd rather wait for his return with making a new release and generating the doi. We'll get back to you mid-January at the latest.
