### 2019-01-07 #1095 @danielhuppmann
@lheagy, the stable doi for the project on zenodo is
[10.5281/zenodo.1470400](https://doi.org/10.5281/zenodo.1470400), the doi for the latest release is [10.5281/zenodo.1491662](https://doi.org/10.5281/zenodo.1491662).

There are a few minor updates since release 0.1.2 (also thanks to issues raised and comments by @Chilipp), but nothing critical, so we'll probably hold off on the next release until we get a few more features implemented.
